Per-table blob size metrics
As it turns out, we don't need to upgrade Dropwizard metrics to v5. We
create the tag manually when constructing the name, and parse it out in
jmx-exporter-config.yaml.

I am able to get the metrics to show up when checking localhost:11001,
but not in the testbox metrics explorer (still trying to figure out what
is going on with my transformer). As mentioned in the Gerrit part of
this PR (link below), we need to actually change our build system to
pull in the GitHub SDK to use this.

https://gerrit.observeinc.com/c/observe/+/53180