Deregister table metric on channel close
We are running into issues where some "ghost" table metrics are being
reported. This happens because we only deregister metrics when the
client is closed, but it is possible we stop sending metrics for a table
before the entire client is closed.

Tested in testbox. At a high-level, just printed out the metric names
with table_blob in them, and made sure they match the ones in the actual
blob.