Flush channels individually if interleaving is disabled
The FlushService tries to flush all channels for the same table, however this
has the downside of causing some channels to be included when they are not
exactly ready. and the next time they do need to flush, the blob size is
smaller than anticipated. To fix that the need for flush is now maintained in
the channel object and the channel cache queries the channels to know which
needs to flush.

The other issue is that by handling the need to flush at the table level, it
leads to race codition where one channel requests flush, while another channel
just flushed and turn the need flush flag off, so the caller waits forever.